refactor: collapse workspace into a single publishable crate
The workspace split (#111) made cargo publish impossible: the layer
crates were path-only dependencies with publish = false, so the root
tzt package could not be uploaded to crates.io.

Bring everything back into one crate while keeping the clean
architecture intact:

- the four layer crates become four top-level modules (src/domain,
  src/usecase, src/infrastructure, src/presentation) with identical
  internal structure; only import paths changed
- per-layer READMEs move next to their modules under src/
- the root Cargo.toml is a plain package again; assert_cmd and
  predicates stay in dev-dependencies, clap keeps the "string"
  feature

What is lost relative to the workspace: the compiler no longer
enforces the dependency direction between layers. The one-way flow
(presentation -> usecase -> domain) is now documented in the READMEs
and guarded by review instead.

cargo publish --dry-run verifies cleanly. All 31 tests pass
unchanged; cognitive complexity is identical to before the move.
