--- crates/infrastructure/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod local_timezone_string_provider;
-mod system_timezone_sources;
-
-pub use local_timezone_string_provider::provide_local_timezone_string;
