--- crates/presentation/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "presentation"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-description = "Presentation layer of tzt: CLI definition, input validation, and output."
-publish = false
-
-[dependencies]
-domain.workspace = true
-usecase.workspace = true
-infrastructure.workspace = true
-clap.workspace = true
-thiserror.workspace = true
