+++ src/domain/README.md
@@ -1,7 +1,7 @@
 # domain
 
 tzt のドメイン層。
-タイムゾーン変換という問題を、型で表現するクレートです。
+タイムゾーン変換という問題を、型で表現するモジュールです。
 
 ## 概要
