+++ src/domain/translation_request.rs
@@ -1,7 +1,7 @@
-use crate::ambiguous_time_strategy::AmbiguousTimeStrategy;
-use crate::conversion_time::ConversionTime;
-use crate::source_timezone::SourceTimezone;
-use crate::target_timezone::TargetTimezone;
+use crate::domain::ambiguous_time_strategy::AmbiguousTimeStrategy;
+use crate::domain::conversion_time::ConversionTime;
+use crate::domain::source_timezone::SourceTimezone;
+use crate::domain::target_timezone::TargetTimezone;
 use chrono::NaiveDateTime;
 use chrono_tz::Tz;
