+++ src/main.rs
@@ -1,3 +1,8 @@
+mod domain;
+mod infrastructure;
+mod presentation;
+mod usecase;
+
 use std::process::ExitCode;
 
 fn main() -> ExitCode {
