+++ src/presentation/command/command_definition.rs
@@ -2,8 +2,8 @@ use super::arguments::{
     ambiguous_time_strategy::ambiguous_time_strategy, from_timezone::from_timezone, time::time,
     to_timezone::to_timezone,
 };
+use crate::infrastructure::provide_local_timezone_string;
 use clap::Command;
-use infrastructure::provide_local_timezone_string;
 
 /// # About:
 /// Provides the command definition for the `tzt` command.
