+++ src/presentation/validator/command_options_validator.rs
@@ -1,8 +1,8 @@
 use super::validation_error::ValidationError;
-use clap::ArgMatches;
-use domain::{
+use crate::domain::{
     AmbiguousTimeStrategy, ConversionTime, SourceTimezone, TargetTimezone, TranslationRequest,
 };
+use clap::ArgMatches;
 
 /// Parses raw CLI strings into a validated `TranslationRequest`.
 ///
