+++ src/usecase/README.md
@@ -16,7 +16,7 @@ DST (夏時間) の切り替わりで、2つの特殊ケースが生まれます
 1. **曖昧な時刻** — DST 終了日は同じ壁時計時刻が2回現れる
 2. **存在しない時刻** — DST 開始日は時計が飛び、間の時刻が消える
 
-この分岐の扱いが、このクレートの存在理由です。
+この分岐の扱いが、このモジュールの存在理由です。
 
 ## 目的
 
@@ -26,7 +26,7 @@ CLI やシステム環境の事情から切り離すこと。
 ## 手法
 
 入力は検証済みの `TranslationRequest` だけを受け取ります。
-このクレートに文字列検証はありません。
+このモジュールに文字列検証はありません。
 
 chrono の `LocalResult` の3分岐をそのまま写し取ります。
